+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Updated dev dependencies.
 - Fixed `<Nav>` and `<Pre>` component `:focus` styles.
+- Use Use `rem` units for scroll shadow styles.
 - Removed `package-lock.json` from `.gitignore` and `.prettierignore`, as it’s disabled in `.npmrc` anyway.
 
 ## 1.0.0